fix: armv7 cross-build type mismatch in thread_affinity
std::min(15LU, processName.size()) fails on armv7-eabihf gcc-15 where
size_type is unsigned int (32-bit) and 15LU is unsigned long (64-bit).
Cast to std::size_t for portability.

+++ core/include/gnuradio-4.0/thread/thread_affinity.hpp
@@ -131,7 +131,7 @@ inline void setProcessName(const std::string_view& processName, int pid = detail
     if (!out.is_open()) {
         throw std::system_error(THREAD_UNINITIALISED, thread_exception(), std::format("setProcessName({},{})", processName, pid));
     }
-    out << std::string{processName.cbegin(), std::min(15LU, processName.size())};
+    out << std::string{processName.cbegin(), std::min(static_cast<std::size_t>(15), processName.size())};
     out.close();
 }
 #else
